% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{load_quote}
\alias{load_quote}
\title{Requesting historic information from Yahoo-Finance}
\usage{
load_quote(symbol = "^GDAXI")
}
\arguments{
\item{symbol}{\code{character}: The Symbol for a specific asset such as stocks,
FX rates, etc.}
}
\description{
Acesses the API of Yahoo Finance to gather close prices of the
selected stock symbol
}
