% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockAnalyzer.R
\name{movingAverage}
\alias{movingAverage}
\title{Using moving average method to profile stock data}
\usage{
movingAverage(data, window, newColname)
}
\arguments{
\item{data}{xts a time series object}

\item{window}{numeric}

\item{newColname}{character}
}
\value{
xts a time series object
}
\description{
Using moving average method to profile stock data
}
\examples{
quantmod::getSymbols("AAPL")
moving_avg_AAPL <- movingAverage(AAPL,300,paste("movingAverage", colnames(AAPL), sep="_"))
}
