% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{select_sample}
\alias{select_sample}
\title{Take a sub-sample}
\usage{
select_sample(df, from, to)
}
\arguments{
\item{df}{\code{data.frame}:}

\item{from}{\code{character}: Starting date in the format "YYYY-MM-DD".}

\item{to}{\code{character}: End date in the format "YYYY-MM-DD".}
}
\description{
uses the output from \link{load_quote} and reduces the full sample to
a sub-sample over the desired timespan.
}
