% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{The definition for the server logic of the app.}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{\code{list}: All objects from the UI that deliver input.}

\item{output}{\code{list}: All objects from the UI that are used by the server
as output.}

\item{session}{\code{list}: All session variables.}
}
\description{
The definition for the server logic of the app.
}
