% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{smooth_sample}
\alias{smooth_sample}
\title{Smoothing}
\usage{
smooth_sample(df, method)
}
\arguments{
\item{df}{\code{data.frame}:}

\item{method}{\code{character}: Various smoothing/detrending/differencing
operations are available.}
}
\description{
Uses the output from \link{select_sample} or \link{load_quote} and creates
an additional column, containing the smoothed/differenced time-series.
}
