% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockAnalyzer.R
\name{summaryStats}
\alias{summaryStats}
\title{Generate summary statistics for profile stock data}
\usage{
summaryStats(data, measurements = colnames(data))
}
\arguments{
\item{data}{xts a time series object}

\item{measurements}{character}
}
\value{
tibble
}
\description{
Generate summary statistics for profile stock data
}
\examples{
quantmod::getSymbols("AAPL")
summary_stats_AAPL <- summaryStats(AAPL)

}
