% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{ui}
\alias{ui}
\title{The UI definition for the App}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\description{
The UI definition for the App
}
\keyword{datasets}
