% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockAnalyzer.R
\name{visExpSmoothing}
\alias{visExpSmoothing}
\title{Using the trend of the stock by using exponential smoothing method}
\usage{
visExpSmoothing(data, alpha, newColname)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr)}

\item{alpha}{A numeric vector of the smoothing parameter which defines the weighting. It should be between 0 and 1}

\item{newColname}{a character vector of the name of the column to be used in moving average calculation (such as Close, Adj Close)}
}
\value{
ggplot line plot of specific stock's historical prices and exponentially smoothed prices
}
\description{
Using the trend of the stock by using exponential smoothing method
}
\examples{
quantmod::getSymbols("AAPL")
visES_AAPL <- visExpSmoothing(AAPL, 0.3, paste("exponentialSmoothing", colnames(AAPL), sep="_"))
}
