% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockAnalyzer.R
\name{summaryStats}
\alias{summaryStats}
\title{Generate summary statistics for profile stock data}
\usage{
summaryStats(data, measurements = c("High", "Low", "Open", "Close"))
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), a lazy data frame (e.g. from dbplyr or dtplyr), or a time series object (xts)}

\item{measurements}{character}
}
\value{
tibble
}
\description{
Generate summary statistics for profile stock data
}
\examples{
quantmod::getSymbols("AAPL")
summary_stats_AAPL <- summaryStats(AAPL, measurements=c("High", "Low", "Open", "Close"))

}
